feat: Update data types for image and peak time in DL1Event and RootDL1Event
- Changed data types from Eigen::VectorXd and RVecD to Eigen::VectorXf and RVecF for image and peak time in DL1Event and RootDL1Event classes.
- Updated related code in RootEventSource and RootWriter to reflect the new data types.
- Adjusted image processing in ImageProcessor to cast images to float type.

+++ include/DL1Event.hh
@@ -23,8 +23,8 @@ class DL1Camera
     DL1Camera& operator=(DL1Camera&& other) noexcept = default;
 
     ImageParameters image_parameters;
-    Eigen::VectorXd image;
-    Eigen::VectorXd peak_time;
+    Eigen::VectorXf image;
+    Eigen::VectorXf peak_time;
     Eigen::Vector<bool, -1> mask;
 };
 class DL1Event: public BaseTelContainer<DL1Camera>
